refactor: product 관련 api 로직 한 파일로 통일
