+++ src/api/products.ts
@@ -1,7 +1,23 @@
 import { api } from "../lib/axios";
-import type { ListProductsData, ListProductsParams } from "../types/products";
+import type {
+  ListProductsData,
+  ListProductsParams,
+  Product,
+} from "../types/products";
 
-// 3) 목록 조회 함수
+// 메인 페이지 신상품 조회
+export async function getNewProducts(): Promise<Product[]> {
+  const res = await api.get("/products/main");
+  return res.data.data.products as Product[];
+}
+
+// 메인 페이지 베스트 상품 조회
+export async function getBestProducts(): Promise<Product[]> {
+  const res = await api.get("/products/main"); // 주문 구현이후 수정
+  return res.data.data.products as Product[];
+}
+
+// 목록 조회 함수
 export async function listProducts(
   params: ListProductsParams,
 ): Promise<ListProductsData> {
